requirements.txt
pandas
numpy
matplotlib
seaborn
dbt
dbt-postgres
psycopg2
telethon
sqlalchemy
databases
scikit-learn
scipy
python-dotenv
xgboost
mlflow
imblearn
tensorflow
shap
lime
mlflow
flask
click
flask_sqlalchemy
Flask-Migrate